Added Relationship between UserCompanyRole-Company-Role

+++ src/company/company.service.ts
@@ -1,18 +1,69 @@
-import { Inject, Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { CreateCompanyDto } from './dto/create-company.dto';
 import { UpdateCompanyDto } from './dto/update-company.dto';
-import { Model } from 'mongoose';
+import { Model, MongooseError } from 'mongoose';
 import { Company } from './entities/company.entity';
 import { InjectModel } from '@nestjs/mongoose';
+import { CompanyUserRoleService } from 'src/company-user-role/company-user-role.service';
+import { CreateCompanyUserRoleDto } from 'src/company-user-role/dto/create-company-user-role.dto';
+import { ValidationError, validateOrReject } from 'class-validator';
+import { RoleService } from 'src/role/role.service';
+import { RoleDocument } from 'src/role/entities/role.entity';
+import { DeleteResult } from 'mongodb';
+import { CompanyUserRoleDocument } from 'src/company-user-role/schemas/company-user-role.schema';
 
 @Injectable()
 export class CompanyService {
   constructor(
     @InjectModel(Company.name) private readonly companyModel: Model<Company>,
+    private readonly companyUserRoleService: CompanyUserRoleService,
+    private readonly roleService: RoleService,
   ) {}
 
-  create(createCompanyDto: CreateCompanyDto) {
+  async create(createCompanyDto: CreateCompanyDto) {
     const newCompany = new this.companyModel(createCompanyDto);
+
+    //Whenever we create a company we also set the Default Role ID to 'Owner'
+    let role: RoleDocument | null;
+    try {
+      role = await this.roleService.findOneByName('owner');
+      if (!role?.id) {
+        throw new MongooseError('Unable to fetch the default Role ID');
+      }
+    } catch {
+      throw new MongooseError('Unable to fetch the default Role ID');
+    }
+
+    //Create the companyUserRoleDTO Object and validate it.
+    const companyUserRole = new CreateCompanyUserRoleDto();
+    companyUserRole.companyId = newCompany.id;
+    companyUserRole.roleId = role?.id;
+    companyUserRole.userId = createCompanyDto.userId;
+    try {
+      await validateOrReject(companyUserRole);
+    } catch (error: any) {
+      throw new BadRequestException(
+        (error as ValidationError[]).flatMap((e) =>
+          Object.values(e.constraints ?? e.toString),
+        ),
+      );
+    }
+
+    //Whenever we create a company we also create an associated CompanyUserRole
+    let createdCompanyUserRole: CompanyUserRoleDocument;
+    try {
+      createdCompanyUserRole =
+        await this.companyUserRoleService.create(companyUserRole);
+    } catch (error) {
+      throw error;
+    }
+
+    if (!createdCompanyUserRole?.id) {
+      throw new MongooseError(
+        'Unable to create an entry for companyUserRole table',
+      );
+    }
+
     return newCompany.save();
   }
 
@@ -31,7 +82,23 @@ export class CompanyService {
   }
 
   //`This action removes a #${id} company`
-  remove(id: string) {
+  async remove(id: string) {
+    //Before Deleting an company delete associated CompanyUserRole as well
+    let deletedCompanyUserRole: DeleteResult;
+    try {
+      deletedCompanyUserRole =
+        await this.companyUserRoleService.removeByCompanyId(id);
+    } catch (err) {
+      console.log(err);
+      throw new MongooseError('Unable to delete associated CompanyUserRole');
+    }
+
+    if (!deletedCompanyUserRole.deletedCount) {
+      //There should be at least one CompanyUserRole record. Because we are creating it by default while creating the company.
+      //If execution is reaching here then there is a bug in the code while creating the company. Please report it to the developer.
+      throw new MongooseError('No associated CompanyUserRole found');
+    }
+
     return this.companyModel
       .findByIdAndDelete(id, { new: true, lean: true })
       .exec();
